chore: Rename `BaseControllerV2` module to `BaseController`
The internal `BaseControllerV2` module has been renamed to
`BaseController`. No external-facing changes. The `V2` in the module
name was an artifact from when we had two controller versions.
